Move a method to a helper file
